+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ------------------------------------------------------------------------
 ### Added
 - Adds ``colorama`` for ``Windows`` systems
+- Overrides ``__repr__``
 
 ### Changed
 - Renames: `` text`` -> ``fore``,  ``background`` -> ``back``